requirements.txt
langchain
langchain-google-genai
langchain-community
tavily-python
firecrawl-py
python-dotenv
requests